SHOPPING CART
Updated Shopping Cart

+++ ShoppingCartBackEnd/src/main/java/com/niit/shoppingcart/dao/CartDAO.java
@@ -14,6 +14,6 @@ public interface CartDAO {
 	
 	public String delete(int id);
 	
-	public int getTotal(String id);
+	public long getTotal(String id);
 
 }
